requirements.txt
flask==2.3.2
PyMySQL==1.1.0
python-dotenv==1.0.0